\name{PMDtest-package}
\alias{PMDtest-package}
\alias{PMDtest}
\docType{package}
\title{
\packageTitle{PMDtest}
}
\description{
\packageDescription{PMDtest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PMDtest}
\packageIndices{PMDtest}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PMDtest}

Maintainer: \packageMaintainer{PMDtest}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
